1.將id的型態改成Long, 2.新增食譜管理系統, 3.增加新增官方食譜(食譜管理系統) 4.修改一般會員的新增食譜及修改食譜功能(部分修改)

+++ src/main/java/com/example/quick_recipe_system/Controller/FavoriteController.java
@@ -52,7 +52,7 @@ public class FavoriteController {
     }
 
     @PostMapping("/favorite/add")
-    public String addFavorite(@RequestParam Integer recipeId,
+    public String addFavorite(@RequestParam Long recipeId,
             HttpSession session,
             RedirectAttributes redirectAttributes,
             HttpServletRequest request) { // 負責抓取使用者是從哪頁點擊的
@@ -94,7 +94,7 @@ public class FavoriteController {
 
     // 3. 處理「取消收藏」的動作
     @PostMapping("/favorite/remove")
-    public String removeFavorite(@RequestParam Integer recipeId,
+    public String removeFavorite(@RequestParam Long recipeId,
             HttpSession session,
             RedirectAttributes redirectAttributes) {
