+++ src/main/java/com/example/quick_recipe_system/repository/FavoriteRepository.java
@@ -10,17 +10,17 @@ import org.springframework.transaction.annotation.Transactional;
 import com.example.quick_recipe_system.entity.Favorite;
 
 @Repository
-public interface FavoriteRepository extends JpaRepository<Favorite, Integer> {
+public interface FavoriteRepository extends JpaRepository<Favorite, Long> {
     // 檢查是否重複
-    boolean existsByUsernameAndRecipeId(String username, Integer recipeId);
+    boolean existsByUsernameAndRecipeId(String username, Long recipeId);
 
     //  新增 1：找出這個人的所有收藏紀錄
     @EntityGraph(attributePaths = {"recipe"})
     List<Favorite> findByUsername(String username);
 
     //  新增 2：根據帳號與食譜 ID 刪除收藏紀錄
-    void deleteByUsernameAndRecipeId(String username, Integer recipeId);
+    void deleteByUsernameAndRecipeId(String username, Long recipeId);
 
     @Transactional // 標記這是一個會更動資料庫的交易動作
-    void deleteByRecipeId(Integer recipeId);
+    void deleteByRecipeId(Long recipeId);
 }
